Remove old ze_result_t WA in negative_tests

+++ negative_tests/core/test_cmdlist_errors/src/test_cmdlist_errors.cpp
@@ -26,18 +26,18 @@ TEST(
 
   descriptor.pNext = nullptr;
   ze_command_list_handle_t cmdlist = nullptr;
-  EXPECT_EQ(uint64_t(ZE_RESULT_ERROR_INVALID_NULL_HANDLE),
-            uint64_t(zeCommandListCreate(lzt::get_default_context(), nullptr,
-                                         &descriptor, &cmdlist)));
+  EXPECT_EQ(ZE_RESULT_ERROR_INVALID_NULL_HANDLE,
+            zeCommandListCreate(lzt::get_default_context(), nullptr,
+                                &descriptor, &cmdlist));
 }
 TEST(
     CommandListCreateNegativeTests,
     GivenInvalidCommandQueueDescriptiorWhileCreatingCommandListThenInvalidNullPointerIsReturned) {
   const ze_device_handle_t device = lzt::zeDevice::get_instance()->get_device();
   ze_command_list_handle_t cmdlist = nullptr;
-  EXPECT_EQ(uint64_t(ZE_RESULT_ERROR_INVALID_NULL_POINTER),
-            uint64_t(zeCommandListCreate(lzt::get_default_context(), device,
-                                         nullptr, &cmdlist)));
+  EXPECT_EQ(ZE_RESULT_ERROR_INVALID_NULL_POINTER,
+            zeCommandListCreate(lzt::get_default_context(), device, nullptr,
+                                &cmdlist));
 }
 TEST(
     CommandListCreateNegativeTests,
@@ -48,9 +48,9 @@ TEST(
   descriptor.stype = ZE_STRUCTURE_TYPE_COMMAND_LIST_DESC;
 
   descriptor.pNext = nullptr;
-  EXPECT_EQ(uint64_t(ZE_RESULT_ERROR_INVALID_NULL_POINTER),
-            uint64_t(zeCommandListCreate(lzt::get_default_context(), device,
-                                         &descriptor, nullptr)));
+  EXPECT_EQ(ZE_RESULT_ERROR_INVALID_NULL_POINTER,
+            zeCommandListCreate(lzt::get_default_context(), device, &descriptor,
+                                nullptr));
 }
 TEST(
     CommandListCreateImmediateNegativeTests,
@@ -63,20 +63,18 @@ TEST(
   descriptor.mode = ZE_COMMAND_QUEUE_MODE_SYNCHRONOUS;
   descriptor.priority = ZE_COMMAND_QUEUE_PRIORITY_NORMAL;
   ze_command_list_handle_t cmdlist_immediate = nullptr;
-  EXPECT_EQ(
-      uint64_t(ZE_RESULT_ERROR_INVALID_NULL_HANDLE),
-      uint64_t(zeCommandListCreateImmediate(lzt::get_default_context(), nullptr,
-                                            &descriptor, &cmdlist_immediate)));
+  EXPECT_EQ(ZE_RESULT_ERROR_INVALID_NULL_HANDLE,
+            zeCommandListCreateImmediate(lzt::get_default_context(), nullptr,
+                                         &descriptor, &cmdlist_immediate));
 }
 TEST(
     CommandListCreateImmediateNegativeTests,
     GivenInvalidCommandQueueDescriptionWhileCreatingCommandListImmediateThenInvalidNullPointerIsReturned) {
   const ze_device_handle_t device = lzt::zeDevice::get_instance()->get_device();
   ze_command_list_handle_t cmdlist_immediate = nullptr;
-  EXPECT_EQ(
-      uint64_t(ZE_RESULT_ERROR_INVALID_NULL_POINTER),
-      uint64_t(zeCommandListCreateImmediate(lzt::get_default_context(), device,
-                                            nullptr, &cmdlist_immediate)));
+  EXPECT_EQ(ZE_RESULT_ERROR_INVALID_NULL_POINTER,
+            zeCommandListCreateImmediate(lzt::get_default_context(), device,
+                                         nullptr, &cmdlist_immediate));
 }
 TEST(
     CommandListCreateImmediateNegativeTests,
@@ -91,9 +89,9 @@ TEST(
   descriptor.mode = ZE_COMMAND_QUEUE_MODE_SYNCHRONOUS;
   descriptor.priority = ZE_COMMAND_QUEUE_PRIORITY_NORMAL;
 
-  EXPECT_EQ(uint64_t(ZE_RESULT_ERROR_INVALID_NULL_POINTER),
-            uint64_t(zeCommandListCreateImmediate(
-                lzt::get_default_context(), device, &descriptor, nullptr)));
+  EXPECT_EQ(ZE_RESULT_ERROR_INVALID_NULL_POINTER,
+            zeCommandListCreateImmediate(lzt::get_default_context(), device,
+                                         &descriptor, nullptr));
 }
 
 } // namespace
